Renamed Package name so that it is consistent with our other Pentaho projects
