+++ src/main/scala/uk/gov/nationalarchives/pdi/test/DatabaseManager.scala
@@ -1,4 +1,4 @@
-package uk.gov.nationalarchives.pentaho
+package uk.gov.nationalarchives.pdi.test
 
 import java.io.File
 import java.sql.{ DriverManager, Statement }
