+++ src/main/scala/uk/gov/nationalarchives/pdi/test/WorkflowManager.scala
@@ -1,4 +1,4 @@
-package uk.gov.nationalarchives.pentaho
+package uk.gov.nationalarchives.pdi.test
 
 import org.pentaho.di.core.KettleEnvironment
 import org.pentaho.di.core.compress.{ CompressionPluginType, NoneCompressionProvider }
